requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.2.1
django-datatables-view==1.17.0
django-mysql==3.0.0.post1
django-oauth-toolkit==1.2.0
djangorestframework==3.9.3
html2text==2018.1.9
idna==2.8
mysqlclient==1.4.2.post1
oauthlib==3.0.1
Pillow==6.0.0
pytz==2019.1
qrcode==6.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.24.3
